refactor: improve project directory structure
